+++ pkg/util/k8sutil/probes/cmd.go
@@ -1,0 +1,41 @@
+//
+// DISCLAIMER
+//
+// Copyright 2026 ArangoDB GmbH, Cologne, Germany
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// Copyright holder is ArangoDB GmbH, Cologne, Germany
+//
+
+package probes
+
+import (
+	core "k8s.io/api/core/v1"
+)
+
+type CMDProbeConfig struct {
+	// Command to be executed
+	Command []string
+
+	Common `json:",inline"`
+}
+
+// Create creates a probe from given config
+func (config CMDProbeConfig) Create() *core.Probe {
+	return config.config(core.ProbeHandler{
+		Exec: &core.ExecAction{
+			Command: config.Command,
+		},
+	})
+}